api-service/requirements.txt
django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pika==1.3.2
python-consul==1.1.0
django-cors-headers==4.4.0
Pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
